Create django app appsutils
'apputils' will house all the models, constants, classes and functions
share with all the other apps.

Also includes a change to settings.py to set BASE_DIR to the django's
project directory and not at the root of the working tree.
